if empty list, should return true

+++ poseidon/controllers/bcf/bcf.py
@@ -337,6 +337,7 @@ class BcfProxy(JsonMixin, CookieAuthControllerProxy):
     def unmirror_mac(self, mac, switch, port, messages=None):
         status = None
         kill_list = self.get_seq_by_mac(mac)
+        if not kill_list: return True
         for kill in kill_list:
             self.logger.debug('unmirroring: {0}'.format(kill))
             self.mirror_traffic(kill, mirror=False)
